Move admin advisor to separate file
```release-note
NONE
```
